[step4] refactor 테스트코드
