+++ src/test/java/racing/race/CarsTest.java
@@ -1,16 +1,12 @@
-package racing;
+package racing.race;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import racing.race.Cars;
-import racing.strategy.MoveNumberStrategy;
-import racing.strategy.RandomNumberStrategy;
-import racing.strategy.StopNumberStrategy;
 
-public class RacingCarTest {
+class CarsTest {
 
   Cars cars;
   private static final int CAR_NUMBER = 3;
@@ -36,44 +32,20 @@ public class RacingCarTest {
         });
   }
 
-  @Test
-  void 멈춤() {
-    StopNumberStrategy stopNumberStrategy = new StopNumberStrategy();
-    cars.getGameCars().get(1).move(stopNumberStrategy);
-
-    assertThat(cars.getGameCars().get(1).moveDistance()).isEqualTo(0);
-  }
-
-  @Test
-  void 전진() {
-    MoveNumberStrategy moveNumberStrategy = new MoveNumberStrategy();
-    cars.getGameCars().get(1).move(moveNumberStrategy);
-
-    assertThat(cars.getGameCars().get(1).moveDistance()).isEqualTo(1);
-  }
-
+  // TODO move()에서 전략을 매개변수로 받지 않고 Car를 생성할 때 이 자동차는 랜덤값으로 전진하는 자동차야. 같은 개념으로 상태로 들고 있게
   @Test
   void 레이싱게임진행() {
     cars.race(TRY_NUMBER);
   }
 
   @Test
-  void 랜덤값입력() {
-    RandomNumberStrategy randomNumberStrategy = new RandomNumberStrategy();
-
-    int result = randomNumberStrategy.getNumber();
-
-    assertThat(result).isBetween(0, 9);
-  }
-
-  @Test
-  void 이동횟수음수입력() {
+  void 이동횟수는음수를입력할수없다() {
     assertThatExceptionOfType(NumberFormatException.class)
         .isThrownBy(() -> cars.race(-1));
   }
 
   @Test
-  void 자동차갯수입력() {
+  void 자동차생성할수있다() {
     assertThat(cars.getGameCars().size()).isEqualTo(CAR_NUMBER);
   }
